- Resolução de bug: nota dos livros que aparecem no perfil.

+++ app/src/main/java/br/edu/ifce/swappers/swappers/dao/BookDAO.java
@@ -4,6 +4,7 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.util.Log;
 
 import java.util.ArrayList;
 
@@ -88,7 +89,7 @@ public class BookDAO {
                 book.setPhoto(cursor.getString(4));
                 book.setNumberPage(cursor.getInt(5));
                 book.setSynopsis(cursor.getString(6));
-                book.setEvaluationAvarage(cursor.getInt(7));
+                book.setEvaluationAvarage(cursor.getFloat(7));
 
                 bookList.add(book);
             }while (cursor.moveToNext());
